chore: rename chain.AnonToken to TokenTx
