+++ internal/chain/token_transaction.go
@@ -1,12 +1,13 @@
 package chain
 
 import (
-	"bytes"
-	"crypto/sha256"
 	"fmt"
 	"math/big"
 
+	"github.com/AccumulateNetwork/accumulated/types/api"
+
 	"github.com/AccumulateNetwork/accumulated/types"
+	types2 "github.com/AccumulateNetwork/accumulated/types/anonaddress"
 	"github.com/AccumulateNetwork/accumulated/types/api/transactions"
 	"github.com/AccumulateNetwork/accumulated/types/state"
 	"github.com/AccumulateNetwork/accumulated/types/synthetic"
@@ -14,169 +15,150 @@ import (
 
 type TokenTx struct{}
 
-func (TokenTx) updateChain() types.ChainType { return types.ChainTypeToken }
+func (TokenTx) Type() types.TxType { return types.TxTypeTokenTx }
 
-func (TokenTx) BeginBlock() {}
+func (c TokenTx) BeginBlock() {}
 
-func (TokenTx) CheckTx(st *state.StateEntry, tx *transactions.GenTransaction) error {
-	if st.AdiState == nil {
-		return fmt.Errorf("identity state does not exist for anonymous transaction")
-	}
-
-	withdrawal := transactions.TokenSend{} //api.TokenTx{}
-	_, err := withdrawal.Unmarshal(tx.Transaction)
-	if err != nil {
-		return fmt.Errorf("error with send token in TokenTransactionValidator.CheckTx, %v", err)
-	}
-
-	tas := &state.TokenAccount{}
-	err = tas.UnmarshalBinary(st.ChainState.Entry)
-	if err != nil {
-		return fmt.Errorf("cannot unmarshal token account, %v", err)
-	}
-	err = canSendTokens(st, tas, &withdrawal)
-	return err
+func (c TokenTx) CheckTx(st *state.StateEntry, tx *transactions.GenTransaction) error {
+	return nil
 }
 
-func (TokenTx) DeliverTx(st *state.StateEntry, tx *transactions.GenTransaction) (*DeliverTxResult, error) {
-	//need to do everything done in "check" and also create a synthetic transaction to add tokens.
-	withdrawal := transactions.TokenSend{} //api.TokenTx{}
-	_, err := withdrawal.Unmarshal(tx.Transaction)
-	if err != nil {
-		return nil, fmt.Errorf("error with send token in TokenTransactionValidator.DeliverTx, %v", err)
-	}
-
-	ids := &state.AdiState{}
-	err = ids.UnmarshalBinary(st.AdiState.Entry)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling adi")
-	}
-
-	tas := &state.TokenAccount{}
-	err = tas.UnmarshalBinary(st.AdiState.Entry)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling token account")
-	}
-
-	err = canSendTokens(st, tas, &withdrawal)
-
-	if ids == nil {
-		return nil, fmt.Errorf("invalid identity state retrieved for token transaction")
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	keyHash := sha256.Sum256(tx.Signature[0].PublicKey)
-	if !ids.VerifyKey(keyHash[:]) {
-		return nil, fmt.Errorf("key not authorized for signing transaction")
-	}
-
-	//need to do a nonce check.
-	if tx.SigInfo.Nonce <= ids.Nonce {
-		return nil, fmt.Errorf("invalid nonce in transaction, cannot proceed")
-	}
-
-	txid := tx.TransactionHash()
-
+func (c TokenTx) DeliverTx(st *state.StateEntry, tx *transactions.GenTransaction) (*DeliverTxResult, error) {
 	res := new(DeliverTxResult)
-	res.Submissions = make([]*transactions.GenTransaction, 0, len(withdrawal.Outputs))
-
-	txAmt := big.NewInt(0)
-	var amt big.Int
-	for _, val := range withdrawal.Outputs {
-		amt.SetUint64(val.Amount)
-
-		//accumulate the total amount of the transaction
-		txAmt.Add(txAmt, &amt)
-
-		//extract the target identity and chain from the url
-		adi, chainPath, err := types.ParseIdentityChainPath(&val.Dest)
-		if err != nil {
-			return nil, err
-		}
-		destUrl := types.String(chainPath)
-
-		//get the identity id from the adi
-		idChain := types.GetIdentityChainFromIdentity(&adi)
-		if idChain == nil {
-			return nil, fmt.Errorf("Invalid identity chain for %s", adi)
-		}
-
-		//populate the synthetic transaction, each submission will be signed by BVC leader and dispatched
-		sub := new(transactions.GenTransaction)
-		res.AddSyntheticTransaction(sub)
-
-		//set the identity chain for the destination
-		sub.Routing = types.GetAddressFromIdentity(&adi)
-
-		//set the chain id for the destination
-		sub.ChainID = types.GetChainIdFromChainPath(&chainPath).Bytes()
-
-		//set the transaction instruction type to a synthetic token deposit
-		depositTx := synthetic.NewTokenTransactionDeposit(txid[:], &tas.ChainUrl, &destUrl)
-		err = depositTx.SetDeposit(&tas.TokenUrl.String, &amt)
-		if err != nil {
-			return nil, fmt.Errorf("unable to set deposit for synthetic token deposit transaction")
-		}
-
-		sub.Transaction, err = depositTx.MarshalBinary()
-		if err != nil {
-			return nil, fmt.Errorf("unable to set sender info for synthetic token deposit transaction")
-		}
-	}
-
-	//subtract the transfer amount from the balance
-	err = tas.SubBalance(txAmt)
-	if err != nil {
-		return nil, err
-	}
-
-	//issue a state change...
-	tasso, err := tas.MarshalBinary()
-
-	if err != nil {
-		return nil, err
-	}
-
-	//return a transaction state object
-	res.AddStateData(st.ChainId, tasso)
-	return res, nil
-}
-
-func canSendTokens(st *state.StateEntry, tas *state.TokenAccount, withdrawal *transactions.TokenSend) error {
-	if st.ChainState == nil {
-		return fmt.Errorf("no account exists for the chain")
-	}
 
 	if st.AdiState == nil {
-		return fmt.Errorf("no identity exists for the chain")
+		return nil, fmt.Errorf("identity state does not exist for anonymous account")
 	}
 
-	if withdrawal == nil {
-		//defensive check / shouldn't get where.
-		return fmt.Errorf("withdrawl account doesn't exist")
+	//now check to make sure this is really an anon account
+	if st.AdiHeader.Type != types.ChainTypeAnonTokenAccount {
+		return nil, fmt.Errorf("account adi is not an anonymous account type")
 	}
 
-	//verify the tx.from is from the same chain
-	fromChainId := types.GetChainIdFromChainPath(&withdrawal.AccountURL)
-	if bytes.Equal(st.ChainId[:], fromChainId[:]) {
-		return fmt.Errorf("from state object transaction account doesn't match transaction")
+	var err error
+	withdrawal := api.TokenTx{}
+	err = withdrawal.UnmarshalBinary(tx.Transaction)
+	if err != nil {
+		return nil, fmt.Errorf("error with send token, %v", err)
+	}
+
+	//need to derive chain id for coin type account.
+	adi, chain, _ := types.ParseIdentityChainPath(withdrawal.From.AsString())
+	if adi != chain {
+		return nil, fmt.Errorf("cannot specify sub accounts for anonymous token chains")
+	}
+	//specify the acme tokenUrl
+	acmeTokenUrl := types.String("dc/ACME")
+
+	//this is the actual account url the acme tokens are being sent from
+	withdrawal.From = types.UrlChain{types.String(adi)}
+
+	//get the ChainId of the acme account for the anon address.
+	accountChainId := types.GetChainIdFromChainPath(withdrawal.From.AsString())
+
+	//because we use a different chain for the anonymous account, we need to fetch it.
+	st.ChainId = accountChainId
+	st.ChainState, err = st.DB.GetCurrentEntry(accountChainId[:])
+	if err != nil {
+		return nil, fmt.Errorf("chain state for account not esablished")
+	}
+
+	tokenAccountState := new(state.TokenAccount)
+	err = tokenAccountState.UnmarshalBinary(st.ChainState.Entry)
+	if err != nil {
+		return nil, err
 	}
 
 	//now check to see if we can transact
 	//really only need to provide one input...
+	//now check to see if the account is good to send tokens from
 	amt := types.Amount{}
-	var outAmt big.Int
-	for _, val := range withdrawal.Outputs {
-		amt.Add(amt.AsBigInt(), outAmt.SetUint64(val.Amount))
+	txAmt := big.NewInt(0)
+	for _, val := range withdrawal.To {
+		amt.Add(amt.AsBigInt(), txAmt.SetUint64(val.Amount))
 	}
 
-	//make sure the user has enough in the account to perform the transaction
-	if tas.GetBalance().Cmp(amt.AsBigInt()) < 0 {
-		///insufficient balance
-		return fmt.Errorf("insufficient balance")
+	if !tokenAccountState.CanTransact(amt.AsBigInt()) {
+		return nil, fmt.Errorf("insufficient balance")
 	}
-	return nil
+
+	//so far, so good.  Now we need to check to make sure the signing address is ok.  maybe look at moving this upstream from here.
+	//if we get to this function we know we at least have 1 signature
+	address := types2.GenerateAcmeAddress(tx.Signature[0].PublicKey)
+
+	//check the addresses to make sure they match
+	if address != string(st.AdiHeader.ChainUrl) {
+		return nil, fmt.Errorf("invalid address, public key address is %s but account %s ", address, st.AdiHeader.ChainUrl)
+	}
+
+	//now build the synthetic transactions.
+	txid := types.Bytes(tx.TransactionHash())
+	for _, val := range withdrawal.To {
+		txAmt.SetUint64(val.Amount)
+		//extract the target identity and chain from the url
+		destAdi, destChainPath, err := types.ParseIdentityChainPath(val.URL.AsString())
+		if err != nil {
+			return nil, err
+		}
+		destUrl := types.String(destChainPath)
+
+		//populate the synthetic transaction, each submission will be signed by BVC leader and dispatched
+		gtx := new(transactions.GenTransaction)
+
+		//set the identity chain for the destination
+		gtx.Routing = types.GetAddressFromIdentity(&destAdi)
+		gtx.ChainID = types.GetChainIdFromChainPath(destUrl.AsString()).Bytes()
+		gtx.SigInfo = new(transactions.SignatureInfo)
+		gtx.SigInfo.URL = destAdi
+
+		depositTx := synthetic.NewTokenTransactionDeposit(txid[:], &st.AdiHeader.ChainUrl, &destUrl)
+		err = depositTx.SetDeposit(&acmeTokenUrl, txAmt)
+		if err != nil {
+			return nil, fmt.Errorf("unable to set deposit for synthetic token deposit transaction, %v", err)
+		}
+
+		gtx.Transaction, err = depositTx.MarshalBinary()
+		if err != nil {
+			return nil, fmt.Errorf("unable to marshal synthetic token transaction deposit, %v", err)
+		}
+
+		res.AddSyntheticTransaction(gtx)
+	}
+
+	err = tokenAccountState.SubBalance(amt.AsBigInt())
+	if err != nil {
+		return nil, fmt.Errorf("error subtracting balance from account acc://%s, %v", st.AdiHeader.ChainUrl, err)
+	}
+
+	txHash := txid.AsBytes32()
+	//create a transaction reference chain acme-xxxxx/0, 1, 2, ... n.
+	//This will reference the txid to keep the history
+	refUrl := fmt.Sprintf("%s/%d", adi, tokenAccountState.TxCount)
+	txr := state.NewTxReference(refUrl, txHash[:])
+	txrData, err := txr.MarshalBinary()
+	if err != nil {
+		return nil, fmt.Errorf("unable to process transaction reference chain %v", err)
+	}
+
+	//now create the chain state object.
+	txRefChain := new(state.Object)
+	txRefChain.Entry = txrData
+	txRefChainId := types.GetChainIdFromChainPath(&refUrl)
+
+	//increment the token transaction count
+	tokenAccountState.TxCount++
+
+	//want to pass back just an interface rather than marshaled data, for now just
+	//do marshaled data
+	data, err := tokenAccountState.MarshalBinary()
+	if err != nil {
+		panic("anon token end block, error marshaling account state.")
+	}
+	st.AdiState.Entry = data
+
+	//now update the state
+	st.DB.AddStateEntry(accountChainId, &txHash, st.AdiState)
+	st.DB.AddStateEntry(txRefChainId, &txHash, txRefChain)
+
+	return res, nil
 }
